@@ -8,78 +8,70 @@
   },
   "license": "GPL-3.0",
   "description": "UI app for Photoview",
-  "dependencies": {
-    "@apollo/client": "^3.5.8",
-    "@babel/preset-typescript": "^7.16.7",
-    "@craco/craco": "^6.4.3",
-    "@headlessui/react": "^1.4.3",
-    "@react-aria/focus": "^3.5.0",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@types/geojson": "^7946.0.8",
-    "@types/jest": "^27.4.0",
-    "@types/mapbox-gl": "^2.6.0",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-helmet": "^6.1.5",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.21",
-    "@types/url-join": "^4.0.1",
-    "autoprefixer": "^9.8.6",
-    "babel-plugin-graphql-tag": "^3.3.0",
-    "blurhash": "^1.1.4",
-    "classnames": "^2.3.1",
-    "connect-history-api-fallback": "^1.6.0",
-    "copy-to-clipboard": "^3.3.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest-dom": "^4.0.1",
-    "fs-extra": "^10.0.0",
-    "i18next": "^21.6.10",
-    "mapbox-gl": "^2.6.1",
-    "postcss": "^7.0.36",
-    "prettier": "^2.5.1",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-blurhash": "^0.1.3",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.25.3",
-    "react-i18next": "^11.15.3",
-    "react-router-dom": "^6.2.1",
-    "react-router-prop-types": "^1.0.5",
-    "react-scripts": "^4.0.3",
-    "react-spring": "^8.0.27",
-    "react-test-renderer": "^17.0.2",
-    "styled-components": "^5.3.3",
-    "subscriptions-transport-ws": "^0.11.0",
-    "tailwind-override": "^0.6.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
-    "typescript": "^4.5.5",
-    "url-join": "^4.0.1"
-  },
   "scripts": {
-    "start": "BROWSER=none PORT=1234 craco start",
-    "build": "craco build",
+    "start": "vite",
+    "build": "vite build",
     "test": "npm run lint && npm run jest -- --watchAll=false",
     "test:ci": "npm run lint && npm run jest:ci",
-    "lint": "npm run lint:types & npm run lint:eslint",
-    "lint:eslint": "eslint ./src --max-warnings 0 --cache --config .eslintrc.js",
-    "lint:types": "tsc --noemit",
+    "lint": "eslint ./src --max-warnings 0 --cache --config .eslintrc.js",
     "jest": "craco test --setupFilesAfterEnv ./testing/setupTests.ts",
     "jest:ci": "CI=true craco test --setupFilesAfterEnv ./testing/setupTests.ts --verbose --ci --coverage",
     "genSchemaTypes": "apollo client:codegen --target=typescript --globalTypesFile=src/__generated__/globalTypes.ts && prettier --write */**/__generated__/*.ts",
     "extractTranslations": "i18next -c i18next-parser.config.js",
     "prepare": "(cd .. && npx husky install)"
   },
+  "dependencies": {
+    "@apollo/client": "^3.6.9",
+    "@babel/preset-typescript": "^7.18.6",
+    "@headlessui/react": "^1.6.6",
+    "@types/geojson": "^7946.0.8",
+    "@types/jest": "^28.1.4",
+    "@types/mapbox-gl": "^2.7.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-helmet": "^6.1.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.25",
+    "@types/url-join": "^4.0.1",
+    "@vitejs/plugin-react": "^1.3.2",
+    "autoprefixer": "^10.4.7",
+    "blurhash": "^1.1.5",
+    "classnames": "^2.3.1",
+    "connect-history-api-fallback": "^2.0.0",
+    "copy-to-clipboard": "^3.3.1",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jest-dom": "^4.0.2",
+    "i18next": "^21.8.13",
+    "mapbox-gl": "^2.9.1",
+    "postcss": "^8.4.14",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-blurhash": "^0.1.3",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-hook-form": "^7.33.1",
+    "react-i18next": "^11.18.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-test-renderer": "^18.2.0",
+    "styled-components": "^5.3.5",
+    "subscriptions-transport-ws": "^0.11.0",
+    "tailwind-override": "^0.6.1",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
+    "typescript": "^4.7.4",
+    "url-join": "^5.0.0",
+    "vite": "^2.9.13",
+    "vite-plugin-svgr": "^2.2.0"
+  },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "apollo": "2.33.9",
-    "apollo-language-server": "1.26.7",
-    "husky": "^7.0.4",
-    "i18next-parser": "^5.4.0",
-    "lint-staged": "^12.3.2",
-    "tsc-files": "1.1.2"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "apollo": "2.34.0",
+    "apollo-language-server": "1.26.9",
+    "husky": "^8.0.1",
+    "i18next-parser": "^6.5.0",
+    "lint-staged": "^13.0.3"
   },
   "overrides": {
     "graphql": "^15.0.0"
