@@ -21,7 +21,7 @@ export function setupLocalization(): void {
     },
 
     react: {
-      useSuspense: process.env.NODE_ENV == 'production',
+      useSuspense: import.meta.env.PROD,
     },
   })
 }
